+++ src/pages/SignUp/index.jsx
@@ -14,6 +14,7 @@ import * as ROUTES from "./../../constants/routes";
 import { Redirect } from "react-router-dom";
 import { SaveFile } from "../../components";
 import { flexbox } from "@material-ui/system";
+import Validator from './../../utility/validator'
 const NodeRSA = require("node-rsa");
 
 const SignUp = () => {
@@ -25,20 +26,7 @@ const SignUp = () => {
   const [loader, setLoader] = React.useState(false);
   const [hash, setHash] = React.useState("");
   const [publichash, SetPublicHash] = React.useState("");
-  function getPassHash() {
-    const tokenString = localStorage.getItem("public_hash");
-    const userToken = JSON.parse(tokenString);
-    return userToken;
-  }
-  function getUserName() {
-    const tokenString = localStorage.getItem("user_name");
-    if(tokenString){
-    const userToken = JSON.parse(tokenString);
-    return userToken;}
-    else{
-      return false;
-    }
-  }
+
   React.useEffect(() => {
     async function setup() {
       await loadWeb3();
@@ -95,21 +83,18 @@ const SignUp = () => {
       setError("Enter a valid username");
     }
   };
-  function getUsername() {
-    const tokenString = localStorage.getItem("user_name");
-    const userToken = JSON.parse(tokenString);
-    return userToken;
-  }
+
   // const EncryptData=()=>{
   //   if()
   //   const userEncryption = await EncrptPublicKey(username, pubKey)
-  // }
+    // }
 
-  const token = getPassHash();
-  const loginUser = getUserName();
+  const token = Validator('publicHash')
+  const loginUser = Validator('username')
   if (token && loginUser) {
     return <Redirect to={ROUTES.DASHBOARD} />;
   }
+
   return (
     <Grid container spacing={4}>
       <Grid item xs={12} sm={12} md={4} lg={4}>
